refactor(orchestration): The `Runner.run_task` method now uses task_spec's arguments
Previously the invocations of this method were weird since they were passing both `task_spec` and `task_spec.arguments`.
The new interface is still sub-optimal and will be changed in the future. The `TaskSpec` is inconvenient to use on the DSL level since it can only hold string values. There needs to be a DSL-level Task object that mirrors the TaskSpec, but can hold passed artifact references.