feat(orchestration): Added support for getting information (size, is_dir, hashes) for a storage location
There can be multiple hashes for a given file/dir calculated by different hashing algorithms. This is done to make it possible to use stronger hash algorithms locally, but also use the weaker built-in algorithms that the cloud storage services use.
The directory hashes are calculated by hashing a JSON-encoded list of file information objects (path, size, hash_name, hash). To preserve the information about empty subdirectories, a subdirectory is represented as zero-sized object with a path that ends with slash (the path for the root dir is an empty string). This is consistent with how the directories are represented in various cloud storage providers (AWS, Azure, GCP). For multiple hashes, multiple directory hashes is calculated as well (based on the corresponding file hashes).
For Google Cloud Storage I use the service-calculated MD5 hash for the individual files. However the documentation states that in some circumstances the returned MD5 hash might be different from locally-computed hash as the service hash might have been calculated based on multiple file chunks.
For local files I use the `hashlib` module to calculate MD5 and SHA256 hashes.