feat(orchestration): Added support for execution caching and reuse
Successful container executions are serialized and put in storage. Before launching a new container task, the system checks whether execution with same cache key already exists in the execution cache. If a matching execution is found in the cache, this execution is reloaded from the cache and used instead of actually execution the task.

Started calculating and attaching data info to all created artifacts (constant data and execution outputs).
Added support for serialization and deserialization to the `_StorageArtifact` and `ContainerExecution` objects.
Added support for getting cache key structure for a `ContainerExecution` object.

There is a lot more to improve here:
* A way to disable or control the cache reuse.
* Cache reuse should be based on artifact content (hash), not lineage.
* Support "latching" to still running executions. Reliably. If same pipeline is executed 10 times in parallel, each task should only be executed once.
* Artifacts should be stored in the DB independently of executions
* Executions should be stored in the DB using UUIDs and the cache entries should just point to the execution IDs.
* Running and failed executions should be stored in the DB.
* Support for "maximum cached data staleness". This controls the cached execution reuse.
* Support for "volatile" components.