feat(orchestration): Support reusing running executions from cache
This feature is the final part of the execution caching support.
Without this feature, running the same long-running component multiple times (with same arguments) would result in multiple executions. Same with long-running task chains. Even if the first few executions are reused, once the second chain caches up, the execution chains would diverge. This is not the case when running executions can be reused from cache. Starting 10 long-running pipelines with differences only in the middle should result in the first part of the pipeline being executed only once.

Implementation:
The implementation is currently pretty naive.
Running executions are put in special in-memory cache from which they can be reused in case there are no Succeeded executions in the execution cache DB.
The running execution reuse respects `max_cached_data_staleness`: the running execution start time should be no older than `max_cached_data_staleness`.
In case multiple running executions are available, the oldest valid one is taken. This ensures the reuse stability. Also the oldest running execution usually finishes sooner.

There might be a small caching race condition: When caching we first check the Succeeded cache and then the Running cache. But the executions progress from Running to Succeeded. So an execution might transition right in the time of the cache check. Fixing this issue would increase the amount of time the cache lock is held.