feat(orchestration): Added attributes to the _StorageArtifact objects that point to the execution that produced the artifact
These attributes are needed to trace the artifact lineage. Both from the Python objects and from the DB entities.

The `_execution` attribute holds the `ContainerExecution` that produced the artifact.
The `_execution_id` attribute holds the DB ID of the `ContainerExecution` that produced the artifact.
The `_output_name` attribute holds the name of the `ContainerExecution` output that produced the artifact.
The `_execution_id` and `_output_name` attributes are currently saved to the DB as part of the `container_executions` table entities. However if the `artifacts` DB table is added, this information will be stored in that table.
The `_execution_id` and `_output_name` attributes are restored when loading the execution from the DB, but the `_execution` attribute is not populated. In the future it can be populated lazily.