feat(orchestration): Added the artifact table to the DB
Individual artifacts are now written to the DB.
Artifacts are currently not read from the DB. This is not needed since `ContainerExecution` has the artifacts. However this can change in the future.

Currently, reused constant artifact data objects do not get `artifact._id`. I should fix this in the future.