chore: Fixed Lightweight Python Components tests so that they properly use the current Python interpreter
This feature got broken when the generator moved to always use shell scripts to launch Python. The tests were using the "python3" found in PATH, not the current interpreter. This especially affects virtual environments and WIndows Python. This commit fixes it.
There are some gotchas this code handles:
* Virtual environments only have "python" executable.
* On Linux systems we cannot just use "python" since in some distros it points to python2.
* On Windows there is no "python3" executable.

So, patching the command-line to use the `sys.executable` path instead of "python3" is more robust.